Make necessary SDK metrics symbols private
Fixes #2389
