+++ opentelemetry-sdk/src/opentelemetry/sdk/_metrics/_internal/sdk_configuration.py
@@ -1,13 +1,15 @@
 from dataclasses import dataclass
-from typing import Sequence
+from typing import TYPE_CHECKING, Sequence
 
 from opentelemetry.sdk._metrics.metric_reader import MetricReader
-from opentelemetry.sdk._metrics.view import View
 from opentelemetry.sdk.resources import Resource
+
+if TYPE_CHECKING:
+    from opentelemetry.sdk._metrics.view import View
 
 
 @dataclass
 class SdkConfiguration:
     resource: Resource
     metric_readers: Sequence[MetricReader]
-    views: Sequence[View]
+    views: Sequence["View"]
