+++ opentelemetry-sdk/tests/metrics/test_measurement_consumer.py
@@ -15,15 +15,20 @@
 from unittest import TestCase
 from unittest.mock import MagicMock, Mock, patch
 
-from opentelemetry.sdk._metrics.measurement_consumer import (
+from opentelemetry.sdk._metrics._internal.measurement_consumer import (
     MeasurementConsumer,
     SynchronousMeasurementConsumer,
 )
+from opentelemetry.sdk._metrics._internal.sdk_configuration import (
+    SdkConfiguration,
+)
 from opentelemetry.sdk._metrics.point import AggregationTemporality
-from opentelemetry.sdk._metrics.sdk_configuration import SdkConfiguration
 
 
-@patch("opentelemetry.sdk._metrics.measurement_consumer.MetricReaderStorage")
+@patch(
+    "opentelemetry.sdk._metrics._internal."
+    "measurement_consumer.MetricReaderStorage"
+)
 class TestSynchronousMeasurementConsumer(TestCase):
     def test_parent(self, _):
